new document loaders refactor to global util
